Move jars code to a separate package
