+++ java/cdk/src/test/java/sleeper/cdk/jars/BuiltJarTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk;
+package sleeper.cdk.jars;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
